refactor: /ppc, added react-query hooks, added new default modal components
