+++ src/app/coordinator/list-areas/update-area-modal-form.tsx
@@ -1,9 +1,7 @@
 "use client";
 
 import React from "react";
-import { Plus } from "lucide-react";
 
-import { Button } from "@/app/_components/ui/button";
 import { Input } from "@/app/_components/ui/input";
 import {
   FormItem,
