+++ src/types/ppc.ts
@@ -2,5 +2,5 @@ import { PPCValues } from "./validation/ppc_form";
 
 export interface PPC extends PPCValues {
   id: string;
-  documentUrl: string
+  documentUrl: string | undefined;
 }
